pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avp-cli"
version = "0.1.0"
description = "Command-line interface for Agent Vault Protocol"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "omar-malek"}
]
keywords = ["avp", "credentials", "security", "ai-agents", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "avp-sdk>=0.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
avp = "avp_cli.main:cli"

[project.urls]
Homepage = "https://github.com/avp-protocol/avp-cli"
Repository = "https://github.com/avp-protocol/avp-cli"
Documentation = "https://avp-protocol.github.io/"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
